requirements.txt
torch
transformers
accelerate
sentencepiece
torchvision
Pillow
opencv-python
pytesseract
huggingface-hub
scipy
